Move lib inside src directory
